+++ comment/api/urls.py
@@ -8,7 +8,12 @@ urlpatterns = [
     path('comments/create/', views.CommentCreate.as_view(), name='comments-create'),
     path('comments/<int:pk>/', views.CommentDetail.as_view(), name='comment-detail'),
     path('comments/<int:pk>/react/<str:reaction>/', views.CommentDetailForReaction.as_view(), name='comments-react'),
-    path('comments/<int:pk>/flag/', views.CommentDetailForFlag.as_view(), name='comments-flag')
+    path('comments/<int:pk>/flag/', views.CommentDetailForFlag.as_view(), name='comments-flag'),
+    path(
+        'comments/<int:pk>/flag/state/change/',
+        views.CommentDetailForFlagStateChange.as_view(),
+        name='comments-flag-state-change'
+    ),
 ]
 
 urlpatterns = format_suffix_patterns(urlpatterns)