+++ comment/api/views.py
@@ -1,11 +1,14 @@
 from django.core.exceptions import ValidationError
 from django.contrib.contenttypes.models import ContentType
 from rest_framework import generics, permissions, status
+from rest_framework.exceptions import PermissionDenied
 from rest_framework.generics import get_object_or_404
 from rest_framework.response import Response
 
 from comment.api.serializers import CommentSerializer, CommentCreateSerializer
-from comment.api.permissions import IsOwnerOrReadOnly, ContentTypePermission, ParentIdPermission, FlagEnabledPermission
+from comment.api.permissions import (
+    IsOwnerOrReadOnly, ContentTypePermission, ParentIdPermission, FlagEnabledPermission, CanChangeFlaggedCommentState
+)
 from comment.models import Comment, Reaction, ReactionInstance, Flag, FlagInstance
 
 
@@ -81,11 +84,42 @@ class CommentDetailForFlag(generics.RetrieveAPIView):
 
     def post(self, request, *args, **kwargs):
         comment = get_object_or_404(Comment, id=kwargs.get('pk'))
-        flag = Flag.objects.get_flag_object(comment)
+        flag = Flag.objects.get_for_comment(comment)
         reason = request.data.get('reason') or request.POST.get('reason')
         info = request.data.get('info') or request.POST.get('info')
         try:
             FlagInstance.objects.set_flag(request.user, flag, reason=reason, info=info)
+        except ValidationError as e:
+            return Response({'error': e.messages}, status=status.HTTP_400_BAD_REQUEST)
+
+        serializer = self.get_serializer(comment)
+        return Response(serializer.data, status=status.HTTP_200_OK)
+
+
+class CommentDetailForFlagStateChange(generics.RetrieveAPIView):
+    queryset = Comment.objects.all()
+    serializer_class = CommentSerializer
+    permission_classes = (CanChangeFlaggedCommentState, )
+
+    def get_serializer_context(self):
+        context = super().get_serializer_context()
+        context['flag_update'] = True
+        return context
+
+    def post(self, request, *args, **kwargs):
+        comment = get_object_or_404(Comment, id=kwargs.get('pk'))
+        flag = Flag.objects.get_for_comment(comment)
+        if not comment.is_flagged:
+            raise PermissionDenied(detail='You do not have permission to perform this action.')
+        state = request.data.get('state') or request.POST.get('state')
+        try:
+            state = flag.get_clean_state(state)
+            if not comment.is_edited and state == flag.RESOLVED:
+                return Response(
+                    {'error': 'The comment must be edited before resolving the flag'},
+                    status=status.HTTP_400_BAD_REQUEST
+                )
+            flag.toggle_state(state, request.user)
         except ValidationError as e:
             return Response({'error': e.messages}, status=status.HTTP_400_BAD_REQUEST)